requirements.txt
eel
nfcpy
jinja2
sounddevice
sqlalchemy
load_dotenv
pymysql
scipy
voicevox-client
